+++ ui/src/components/settings/PageImportCollections.svelte
@@ -430,7 +430,13 @@
     </div>
 </PageWrapper>
 
-<ImportPopup bind:this={importPopup} on:submit={clear} />
+<ImportPopup
+    bind:this={importPopup}
+    on:submit={() => {
+        clear();
+        loadOldCollections();
+    }}
+/>
 
 <style>
     .list-label {